+++ selfdrive/loggerd/encoder.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef ENCODER_H
+#define ENCODER_H
 
 #include <stdio.h>
 #include <stdint.h>
@@ -80,4 +81,6 @@ void encoder_destroy(EncoderState *s);
 
 #ifdef __cplusplus
 }
+#endif
+
 #endif
