+++ selfdrive/loggerd/logger.h
@@ -1,24 +1,25 @@
 #pragma once
 #include <assert.h>
 #include <bzlib.h>
+
 #include <memory>
 #include <mutex>
-// #include <vector>
-#include "capnp/serialize.h"
-#include "common/swaglog.h"
-#include "common/utilpp.h"
+
 #include "common/params.h"
+#include "common/swaglog.h"
 #include "common/util.h"
+#include "common/utilpp.h"
+#include "messaging.hpp"
 
 class LoggerHandle {
- public:
+public:
   LoggerHandle() = default;
   ~LoggerHandle() { close(); }
   void write(uint8_t* data, size_t data_size, bool in_qlog = false);
-  void write(capnp::MessageBuilder& msg, bool in_qlog = false);
+  void write(MessageBuilder& msg, bool in_qlog = false);
 
- private:
-  bool open(const std::string& segment_path, const std::string& log_name, int part, bool has_qlog);
+private:
+  bool open(const std::string& segment_path, const std::string& log_name, bool has_qlog);
   void close();
   std::mutex mutex;
   std::string lock_path;
@@ -30,16 +31,16 @@ class LoggerHandle {
 };
 
 class Logger {
- public:
-  Logger(const char* log_name, bool has_qlog);
-  ~Logger();
-  std::shared_ptr<LoggerHandle> openNext(const char* root_path);
+public:
+  Logger(const char* root_path, const char* log_name, bool has_qlog);
+  ~Logger(){};
+  std::shared_ptr<LoggerHandle> openNext();
   inline std::shared_ptr<LoggerHandle> getHandle() const { return cur_handle; }
   inline int getPart() const { return part; }
   inline const char* getSegmentPath() const { return segment_path.c_str(); }
 
  private:
-  std::string segment_path, log_name;
+  std::string segment_path, log_name, root_path;
   char route_name[64] = {};
   int part = 0;
   bool has_qlog = false;
