+++ selfdrive/loggerd/raw_logger.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef FFV1LOGGER_H
+#define FFV1LOGGER_H
 
 #include <cstdio>
 #include <cstdlib>
@@ -38,3 +39,5 @@ private:
 
   AVFrame *frame = NULL;
 };
+
+#endif
